src/postgresql_training/requirements.txt
psycopg2-binary==2.9.5
scikit-learn==1.1.3
pandas==1.5.1
